refactor(fans): drop re-export indirection in useArtistFans hook
