+++ components/Fans/FansList.tsx
@@ -1,4 +1,4 @@
-import { type Social } from "@/hooks/useArtistFans";
+import { type Social } from "@/types/ArtistSocials";
 
 interface FansListProps {
   fans: Social[];
